pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "hbchen", email = "hbchen7@126.com"},
]
dependencies = ["fastapi>=0.115.12", "uvicorn>=0.34.0", "python-multipart>=0.0.20", "python-dotenv>=1.1.0", "tortoise-orm>=0.24.2", "asyncpg>=0.30.0", "aerich>=0.8.2", "tomlkit>=0.13.2", "aiomysql>=0.2.0", "motor>=3.7.0", "langchain>=0.3.21", "langserve[all]>=0.3.1", "langchain-community>=0.3.20", "pypdf>=5.4.0", "langchain-chroma>=0.2.2", "langchain-openai>=0.3.11", "langchain-ollama>=0.3.0", "autopep8>=2.3.2", "unstructured>=0.17.2", "python-jose[cryptography]>=3.4.0", "passlib[bcrypt]>=1.7.4", "beanie>=1.29.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.aerich]
tortoise_orm = "settings.PostgreSQLConfig"
location = "./migrations"
src_folder = "./."
